+++ src/request.ts
@@ -1,35 +1,38 @@
 import { toUrlParams } from './param';
-
 /** requestAPI的第二个参数类型 */
 export interface IRequestOptions {
     /** GET请求时传递的参数 */
     params?: Record<string, any>;
     /** POST请求时传递的参数 */
     data?: Record<string, any>;
+    /** 采用formData模式请求，可以传入普通的对象，也可以传入标准的formData */
+    formData?: Record<string, any>;
     /** postJson请求时传递的参数 */
     json?: Record<string, any>;
     /** API前缀，不传默认为空 */
     baseUrl?: string;
-    /** 请求方法，一般情况下会自动处理无需手动传 */
+    /** 请求方法，绝大部分情况无需手动指定 */
     method?: string;
     /** headers */
     headers?: Record<string, any>;
-    /** 其它自定义fetchOptions */
+    /** 其它自定义fetchOptions，优先级高于上面所有参数 */
     fetchOptions?: RequestInit;
     /** 请求完成之后触发的钩子，无论成功与否均会触发 */
     afterRequest?: (success: boolean, resp?: any) => void;
-    /** 对响应进行自定义格式化处理 */
+    /** 判断接口是否调用成功，默认规则 resp => resp.code == 0 || resp.code == 200 */
+    checkSuccess?: (resp: any) => boolean;
+    /** 对响应进行自定义格式化处理，包含 checkSuccess 全部能力 */
     responseConverter?: (resp: any) => any;
-    /** 后端未返回 message 时的默认异常文案 */
+    /** 后端未返回 message 时的兜底异常文案 */
     errorMessage?: string;
     /** 发生异常时的处理方法，一般不太建议重写此方法 */
     errorHandler?: (message: string, resp?: any) => void;
     /** 自定义toast实现，为了和UI解耦，方法默认不包含UI处理代码 */
     toastHandler?: (message: string) => void;
-    /** 判断接口是否调用成功，默认规则 resp => resp.code == 0 || resp.code == 200 */
-    checkSuccess?: (resp: any) => boolean;
     /** 是否关闭默认的异常toast */
     silent?: boolean;
+    /** 自定义 fetch.then() ，非常底层的一个方法，非必要请勿使用 */
+    onFetchResponse?: (response: Response) => any;
 }
 
 /**
@@ -56,6 +59,21 @@ export const request = async <T = any>(url: string, options?: IRequestOptions) =
         // eslint-disable-next-line eqeqeq
         checkSuccess: (resp) => resp?.code == 0 || resp?.code == 200,
         toastHandler: (msg: string) => console.error(`[您还没有配置toastHandler，请根据您的UI组件库配置合适的提示方法] ${msg || ''}`),
+        onFetchResponse: (response) => {
+            if (response.status === 200) {
+                return response.json();
+            }
+            const errorResponse = {
+                code: response.status,
+                message: response.statusText,
+                headers: response.headers,
+                // 返回整个response备用，但是一般不推荐用户直接使用
+                _response: response,
+            };
+            const error: any = new Error(errorResponse.message);
+            error.response = errorResponse;
+            throw error;
+        },
     };
     options = Object.assign({}, defaultOptions, options || {});
     // 防止对象引用产生严重bug
@@ -63,6 +81,7 @@ export const request = async <T = any>(url: string, options?: IRequestOptions) =
     const {
         params,
         data,
+        formData,
         json,
         method,
         headers,
@@ -73,9 +92,10 @@ export const request = async <T = any>(url: string, options?: IRequestOptions) =
         errorHandler,
         errorMessage,
         responseConverter,
+        onFetchResponse,
     } = options || {};
-    // headers 优先级高于 fetchOptions里面的headers
-    fetchOptions.headers = { ...(fetchOptions.headers || {}), ...(headers || {}) };
+    // fetchOptions 里面的 headers 优先级高于外部的 headers
+    fetchOptions.headers = { ...(headers || {}), ...(fetchOptions.headers || {}) };
     if (params) {
         fetchOptions.method = method || 'GET';
         url = `${url}${url.indexOf('?') >= 0 ? '&' : '?'}${toUrlParams(params)}`;
@@ -89,6 +109,18 @@ export const request = async <T = any>(url: string, options?: IRequestOptions) =
                 ...(fetchOptions.headers || {}),
             },
         });
+    } else if (formData) {
+        let body = formData;
+        if (!(formData instanceof FormData)) {
+            body = new FormData();
+            for (const key in formData) {
+                body.append(key, formData[key]);
+            }
+        }
+        Object.assign(fetchOptions, {
+            method: method || 'POST',
+            body,
+        });
     } else if (json) {
         Object.assign(fetchOptions, {
             method: method || 'POST',
@@ -101,11 +133,11 @@ export const request = async <T = any>(url: string, options?: IRequestOptions) =
     }
     let resp = null;
     try {
-        resp = await fetch(`${baseUrl || ''}${url}`, fetchOptions).then((res: any) => res.json());
+        resp = await fetch(`${baseUrl || ''}${url}`, fetchOptions).then(onFetchResponse);
     } catch (e: any) {
         console.error(e);
         afterRequest?.(false, { message: e?.message });
-        errorHandler?.(errorMessage || '');
+        errorHandler?.(e.response?.message || errorMessage || '', e.response);
         return undefined;
     }
     resp = responseConverter ? responseConverter(resp) : resp;